allow user to input bond characteristics

+++ src/com/company/RiskyBond.java
@@ -2,6 +2,6 @@ package com.company;
 
 import java.util.HashMap;
 
-public record RiskyBond(int yearsToMaturity, int parValue, int couponRateInPercent, double recoveryRate,
+public record RiskyBond(int yearsToMaturity, int parValue, double couponRateInPercent, double recoveryRate,
                         double hazardRate, HashMap<Integer,Double> lossesGivenDefault) {
 }
